requirements.txt
alabaster==0.7.12
Babel==2.7.0
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
cycler==0.10.0
docutils==0.15.2
idna==2.8
imagesize==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
numpy==1.16.4
packaging==19.2
pandas==0.25.0
paramiko==2.6.0
pycparser==2.19
Pygments==2.5.2
PyNaCl==1.3.0
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
snowballstemmer==2.0.0
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.5
urllib3==1.25.7
xlrd==1.2.0
XlsxWriter==1.1.8
